Implemented MySQL and fixed a Messages bug
The bug was that only the first reported error message was being
displayed. This is intended behavior for the toolset, as it will only
report the highest-priority message (error, then warning, then pass) at
the aggregated output results, but not the intended behavior in the
verifier which should report *all* problems and let the toolset sort it
out later.